refactor(core): remove legacy repository interfaces

+++ src/Application/useCase/Hub/AddHubUseCase.ts
@@ -1,6 +1,6 @@
 import { inject, injectable } from "tsyringe";
-import { IHubTempRepository } from "../../../Infrastructure/Interface/repositories_interfaces/hubRepositories_Interfaces/hubTemp.repository";
-import { IHubRepository } from "../../../Infrastructure/Interface/repositories_interfaces/hubRepositories_Interfaces/hub.repository";
+import { IHubTempRepository } from "../../interfaces/repositories_interfaces/hubRepositories_Interfaces/hubTemp.repository";
+import { IHubRepository } from "../../interfaces/repositories_interfaces/hubRepositories_Interfaces/hub.repository";
 import { IPasswordService } from "../../interfaces/services_Interfaces/password-service.interface";
 import { IMailService } from "../../interfaces/services_Interfaces/email-service.interface";
 import { AppError } from "../../../Domain/utils/customError";
@@ -11,6 +11,7 @@ import { ENV } from "../../../Infrastructure/constants/env";
 import { HUB_MESSAGES } from "../../../Infrastructure/constants/messages/hubMessage";
 import { STATUS } from "../../../Infrastructure/constants/statusCodes";
 
+let addHubCount = 0;
 
 @injectable()
 export class AddHubUseCase implements IAddHubUseCase {
@@ -47,8 +48,15 @@ export class AddHubUseCase implements IAddHubUseCase {
         tempHub.location_lng = extraData.location_lng;
 
         const hubEntity = HubMapper.toCreateHub(tempHub, hashedPassword, imageUrl);
+        if(addHubCount>2) throw new AppError("can only add 2 hub a day");
+
 
         const savedHub = await this._hubRepo.save(hubEntity);
+        if(savedHub){
+            addHubCount++
+            console.log(addHubCount,'addHubCount')
+        };
+
 
         await this._hubTempRepo.delete({ _id: tempHubId });
         if (ENV.IS_PROD) await this._mailService.sendCustomPassword(tempHub.email);
